move labeling_neighbors_helper to preprocessing
